refactor: update folder structure in pages
