+++ src/pages/Run/Run.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
 import { describe, expect, test } from 'vitest'
 
-import { Run } from './index'
+import { Run } from './Run'
 
 describe('Run', () => {
   test.skip('renders the Run view', () => {
