--- src/pages/Run/components/RunConfigurationPanel/components/HeadersForm/components/index.ts
@@ -1,1 +1,0 @@
-export { HeaderInput } from './HeaderInput'
