--- src/pages/Run/components/RunConfigurationPanel/components/index.ts
@@ -1,2 +1,0 @@
-export { HeadersForm } from './HeadersForm'
-export { TestsForm } from './TestsForm'
