--- src/pages/Run/components/index.ts
@@ -1,2 +1,0 @@
-export { RunConfigurationPanel } from './RunConfigurationPanel'
-export { RunResultDisplay } from './RunResultDisplay'
