+++ src/pages/Run/index.ts
@@ -1,0 +1,1 @@
+export { Run } from './Run'
