+++ src/pages/Run/internal/Configuration/TestsConfiguration/index.ts
@@ -1,0 +1,1 @@
+export { TestsConfiguration } from './TestsConfiguration'
