+++ src/pages/Run/internal/Configuration/TestsConfiguration/internal/TestsConfiguration.module.scss
@@ -1,5 +1,5 @@
 @use 'sass:map';
-@use '../../../../../../../style/variables/colors.scss';
+@use '../../../../../../style/variables/colors.scss';
 
 .run-configuration-panel-tests {
   &__icon {
