+++ src/pages/Run/internal/ResultDisplay/index.ts
@@ -1,0 +1,1 @@
+export { ResultDisplay } from './ResultDisplay'
